# sort: Fast path for already sorted data