@Gankro 2015-11-07
Great wins on big runs! :+1: 

I'm a bit confused with what's happening in the small input cases. You mention jemalloc but I don't follow why that would have changed between runs?