@bluss 2015-11-07
> I'm comparing my nightly rustc vs the patched rustc built from the tree.

It's this.