@huonw 2015-11-07
>> I'm comparing my nightly rustc vs the patched rustc built from the tree.

> It's this.

I don't understand why that's relevant? If optimisations are enabled, they should be generating essentially the same code?