@huonw 2015-11-07
Oh, I see, if the patched rustc was built without optimisations, then jemalloc may've been too, and hence runs slower... This is somewhat suspicious, so r=me, except for nailing down the exact cause of the regression (and/or getting a version of the benchmarks that shows it doesn't exist for a "real" build).