@bluss 2015-11-13
Ok, I ran the in tree benchmarks:

```
before

test slice::bench::sort_big_random_large                        ... bench:   4,749,545 ns/iter (+/- 948,440) = 67 MB/s
test slice::bench::sort_big_random_medium                       ... bench:      36,870 ns/iter (+/- 812) = 86 MB/s
test slice::bench::sort_big_random_small                        ... bench:       1,798 ns/iter (+/- 100) = 88 MB/s
test slice::bench::sort_big_sorted                              ... bench:     772,643 ns/iter (+/- 11,515) = 414 MB/s
test slice::bench::sort_random_large                            ... bench:   1,806,359 ns/iter (+/- 36,855) = 44 MB/s
test slice::bench::sort_random_medium                           ... bench:      14,509 ns/iter (+/- 198) = 55 MB/s
test slice::bench::sort_random_small                            ... bench:         685 ns/iter (+/- 20) = 58 MB/s
test slice::bench::sort_sorted                                  ... bench:     397,642 ns/iter (+/- 1,082) = 100 MB/s

after

test slice::bench::sort_big_random_large                        ... bench:   4,718,159 ns/iter (+/- 41,542) = 67 MB/s
test slice::bench::sort_big_random_medium                       ... bench:      36,909 ns/iter (+/- 292) = 86 MB/s
test slice::bench::sort_big_random_small                        ... bench:       1,888 ns/iter (+/- 13) = 84 MB/s
test slice::bench::sort_big_sorted                              ... bench:     615,455 ns/iter (+/- 8,196) = 519 MB/s
test slice::bench::sort_random_large                            ... bench:   1,796,680 ns/iter (+/- 13,319) = 44 MB/s
test slice::bench::sort_random_medium                           ... bench:      14,588 ns/iter (+/- 352) = 54 MB/s
test slice::bench::sort_random_small                            ... bench:         699 ns/iter (+/- 9) = 57 MB/s
test slice::bench::sort_sorted                                  ... bench:     134,979 ns/iter (+/- 1,101) = 296 MB/s
```

Which I think looks good?