@bluss 2015-11-13
The "big" cases sort elements of `type BigSortable = (u64, u64, u64, u64); `, small cases u64, which makes a surprising difference.